Move constants to simulation
